Split tests by files
