+++ api/src/test/kotlin/io/spine/protodata/FieldExtsSpec.kt
@@ -1,0 +1,75 @@
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.protodata
+
+import com.google.protobuf.Empty
+import io.kotest.matchers.shouldBe
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
+import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
+
+@TestInstance(PER_CLASS)
+@DisplayName("`Field` extensions should")
+internal class FieldExtsSpec {
+
+    @Nested
+    inner class
+    `Check if a field is` {
+
+        @Test
+        fun `repeated if list`() {
+            val field = Field.newBuilder()
+                .setList(Empty.getDefaultInstance())
+                .buildPartial()
+
+            field.isRepeated shouldBe true
+        }
+
+        @Test
+        fun `repeated if map`() {
+            val field = Field.newBuilder()
+                .setMap(
+                    Field.OfMap.newBuilder()
+                        .setKeyType(PrimitiveType.TYPE_STRING)
+                        .build())
+                .buildPartial()
+
+            field.isRepeated shouldBe true
+        }
+
+        @Test
+        fun `not repeated`() {
+            val field = Field.newBuilder()
+                .setSingle(Empty.getDefaultInstance())
+                .buildPartial()
+
+            field.isRepeated shouldBe false
+        }
+    }
+}
